+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [1.0.7](https://github.com/atanuroy22/morphe-patches/compare/v1.0.6...v1.0.7) (2026-03-26)
 
+### Bug Fixes
+
+* set external downloader action button default to disabled
+* set override YouTube Music package default to app.revanced.android.apps.youtube.music
+
 ## [1.0.6](https://github.com/atanuroy22/morphe-patches/compare/v1.0.5...v1.0.6) (2026-03-26)
 
 ## [1.0.5](https://github.com/atanuroy22/morphe-patches/compare/v1.0.4...v1.0.5) (2026-03-27)